fix(supervisor): crash-safe activation journal + worker version probe (R9R-004, R9R-001, R9R-005)
R9R-004: make activation crash-safe across the two renames. `activate_binary`
writes a durable activation journal naming active/.prev/staged BEFORE the
renames and clears it after; `process::run` reconciles the journal at startup
(and on a spawn failure) so a crash between the renames recovers to a consistent
`active` binary instead of livelocking on a missing one.

R9R-001: tie a Worker-component activation's observed_version to a real probe of
the on-disk binary (`--version`), not the request label, so bytes that report a
different version fail the health gate and roll back.

R9R-005: split the 942-line activate.rs into `activate/{mod,journal,test_support,
tests}.rs`, all under the 600-line cap, tests moving with their code.
